requirements.txt
Django==1.11.10
argparse==1.2.1
chardet==3.0.4
colorama==0.2.5
dj-database-url==0.5.0
enum34==1.0.4
html5lib==0.999
pathlib==1.0.1
pyScss==1.3.3
pytz==2018.4
requests==2.2.1
six==1.5.2
urllib3==1.7.1
virtualenv==1.11.6
wsgiref==0.1.2
dj-static==0.0.6
static3==0.6.1
gunicorn==19.3.0
path.py==8.1.2
ptyprocess==0.5
static3==0.6.1
psycopg2==2.7.5
django-taggit==0.23.0
